+++ doc/changelog.rst
@@ -68,7 +68,10 @@ Front
 Xtheme
 ~~~~~~
 
+- Enhance default text plugin editor to remarkable markdown editor
 - Add generic snippets plugin for doing simple integrations
+- Add a plugin for displaying category links on shop front
+- Add a linkable image plugin
 
 Classic Gray Theme
 ~~~~~~~~~~~~~~~~~~